JWT Admin Privilege

+++ resources/item.py
@@ -2,7 +2,7 @@ import uuid
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import  SQLAlchemyError
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 
 from schemas import ItemSchema, ItemUpdateSchema
 from models import ItemModel, StoreModel
@@ -20,6 +20,10 @@ class Item(MethodView):
     
     @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege is required.")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -30,6 +34,10 @@ class Item(MethodView):
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
     def put(self, item_data, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege is required.")
+
         item = ItemModel.query.get(item_id)
         if item:
             item.price = item_data["price"]
@@ -53,6 +61,10 @@ class ItemList(MethodView):
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege is required.")
+
         item = ItemModel(**item_data)
         try:
             db.session.add(item)
